+++ src/ladb_opencutlist/ruby/worker/cutlist/cutlist_reset_prices_worker.rb
@@ -1,0 +1,34 @@
+module Ladb::OpenCutList
+
+  require_relative '../../model/attributes/definition_attributes'
+
+  class CutlistResetPricesWorker
+
+    # -----
+
+    def run
+
+      model = Sketchup.active_model
+      return { :errors => [ 'tab.cutlist.error.no_model' ] } unless model
+
+      definitions = model ? model.definitions : nil
+
+      if definitions
+        definitions.each { |definition|
+
+          definition_attributes = DefinitionAttributes.new(definition)
+
+          puts "definition_attributes.price = #{definition_attributes.price}"
+
+          definition_attributes.price = nil
+          definition_attributes.write_to_attributes
+
+        }
+
+      end
+
+    end
+
+  end
+
+end